# backend/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TEAM WORKS API",
    "description": "팀 단위 캘린더 기반 일정관리 + 채팅 애플리케이션 API. API Route 소스 위치: backend/app/api/. v1.7.0 — Postits/Notices/Projects/ProjectSchedules/SubSchedules/Me/Refresh paths 보강",
    "version": "1.7.1",
    "x-source-directory": "backend/app/api"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base URL (Next.js API Routes — backend/app/api/)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "에러 메시지"
          }
        }
      },
...
      "Notice": {
        "type": "object",
        "required": [
          "id",
          "teamId",
          "senderId",
          "senderName",
          "content",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c9d0e1f2-a3b4-5678-cdef-ab9012345678"
          },
          "teamId": {
            "type": "string",
            "format": "uuid",
            "example": "b1c2d3e4-f5a6-7890-bcde-fa1234567890"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "프로젝트 공지일 경우 소속 프로젝트 UUID, 팀 일자별이면 null",
            "example": null
          },
          "senderId": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "senderName": {
            "type": "string",
            "example": "홍길동"
          },
          "content": {
            "type": "string",
            "example": "다음 주 월요일은 휴무입니다."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-20T01:00:00.000Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/teams/{teamId}/projects/{projectId}/notices": {
      "get": {
        "tags": [
          "ProjectNotices"
        ],
        "summary": "프로젝트 공지 목록 조회",
        "description": "프로젝트 전용 공지사항 목록을 createdAt 오름차순으로 조회합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "조회할 프로젝트 UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "공지 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "projectId",
                    "notices"
                  ],
                  "properties": {
                    "projectId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "e1f2a3b4-c5d6-7890-efab-cd1234567890"
                    },
                    "notices": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notice"
                      },
                      "description": "createdAt 오름차순 정렬"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "해당 팀에 접근 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "프로젝트 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "프로젝트를 찾을 수 없습니다."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProjectNotices"
        ],
        "summary": "프로젝트 공지 등록",
        "description": "프로젝트 전용 공지사항을 등록합니다. 팀 구성원이면 누구나 작성할 수 있습니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "공지를 등록할 프로젝트 UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "maxLength": 2000,
                    "example": "이번 주 금요일까지 개발 완료해주세요."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "공지 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notice"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_content": {
                    "value": {
                      "error": "내용은 필수입니다."
                    }
                  },
                  "content_too_long": {
                    "value": {
                      "error": "내용은 최대 2000자까지 입력 가능합니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "해당 팀에 접근 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "프로젝트 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "프로젝트를 찾을 수 없습니다."
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}